Removed autouse fixtures from conftest
